fix test and "refactor" README writing (tests failing)

+++ nipoppy/workflows/dataset_init.py
@@ -3,8 +3,6 @@
 from pathlib import Path
 from typing import Optional
 
-import httpx
-
 try:
     from nipoppy._version import __version__
 except ImportError:
@@ -109,7 +107,9 @@ class InitWorkflow(BaseDatasetWorkflow):
 
         # create directories
         fileops.mkdir(self.dpath_root / NIPOPPY_DIR_NAME, dry_run=self.dry_run)
-        for dpath in self.study.layout.get_paths(directory=True, include_optional=True):
+        for dpath in self.study.layout._get_paths(
+            directory=True, include_optional=True
+        ):
             if self.bids_source is not None and dpath == self.study.layout.dpath_bids:
                 self._handle_bids_source()
             elif (
@@ -119,8 +119,9 @@ class InitWorkflow(BaseDatasetWorkflow):
                 self._handle_container_store()
             else:
                 fileops.mkdir(dpath, dry_run=self.dry_run)
-
-        self._write_readmes()
+                self._write_readme(
+                    dpath, self.study.layout._dpath_descriptions[str(dpath)]
+                )
 
         # create empty pipeline config subdirectories
         for pipeline_type in PipelineTypeEnum:
@@ -225,33 +226,12 @@ class InitWorkflow(BaseDatasetWorkflow):
             dry_run=self.dry_run,
         )
 
-    def _write_readmes(self) -> None:
-        if self.dry_run:
-            return None
-        for dpath, description in self.study.layout.dpath_descriptions:
-            fpath_readme = dpath / self.fname_readme
-            if description is None:
-                continue
-            if dpath.stem != "bids" or self.bids_source is None:
-                fpath_readme.write_text(f"{description}\n")
-            elif self.bids_source is not None and not fpath_readme.exists():
-                gh_org = "bids-standard"
-                gh_repo = "bids-starter-kit"
-                commit = "f2328c58238bdf2088bc587b0eb4198131d8ffe2"
-                path = "templates/README.MD"
-                url = (
-                    "https://raw.githubusercontent.com/"
-                    f"{gh_org}/{gh_repo}/{commit}/{path}"
-                )
-                response = httpx.get(url)
-                readme_content = response.content.decode("utf-8")
-                try:
-                    fpath_readme.write_text(readme_content)
-                except PermissionError:
-                    logger.warning(
-                        f"Permission denied when writing {fpath_readme}. "
-                        "Skipping README creation."
-                    )
+    def _write_readme(self, dpath, description) -> None:
+        if self.dry_run or description is None:
+            return
+
+        fpath_readme = dpath / self.fname_readme
+        fpath_readme.write_text(f"{description}\n")
 
     def _init_manifest_from_bids_dataset(self) -> None:
         """Assume a BIDS dataset with session level folders.
